separate files for metadada, revisions and recycle
